refactor(migration): Move migrate logic to own file
